Added Student Class

+++ Project 1/src/Person/Person.cpp
@@ -1,15 +1,16 @@
 #include "Person.h"
 using namespace std;
+
 const string &Person::getName()  {
         return name;
 }
 void Person::setName(const string &name) {
         Person::name = name;
 }
-uint32_t Person::getAge()  {
+uint8_t Person::getAge()  {
         return age;
 }
-void Person::setAge(uint32_t age) {
+void Person::setAge(uint8_t age) {
         Person::age = age;
 }
 Person::Person() {
@@ -20,6 +21,6 @@ Person::Person(std::string namePar, uint32_t agePar) {
 }
 ostream& operator<<(ostream& os, const Person &p)
 {
-        os << "Name: " << p.name << "\nAge: " << p.age <<endl;
+        os << "Name: " << p.name << "\nAge: " << unsigned(p.age) <<endl;
         return os;
 }
